#18798

Related work items: #18798

+++ src/CPG.Infrastructure.Persistence/DbContexts/ReadModels/CompanyPaymentMethodsReadModel.cs
@@ -3,7 +3,7 @@
 public class CompanyPaymentMethodsReadModel
 {
     public long Id { get; set; }
-    public short MethodType { get; set; }
+    public byte MethodType { get; set; }
     public long CompanyId { get; set; }
     public CompanyReadModel Company { get; set; }
 }
